Refactoring mgmt auth configurations
Uses json format
adds example
updates filters and required classes
updates tests

+++ itest/ldap-test/src/main/java/org/openstack/atlas/api/mgmt/helpers/LDAPTools/JsonConfig.java
@@ -94,7 +94,7 @@ public class JsonConfig {
             conf.setPassword(tmpStr);
 
             tmpLong = (Long) json.get("port");
-            mossoAuthConfig.setPort(tmpLong.intValue());
+            mossoAuthConfig.setPort(tmpLong);
 
             tmpStr = (String) json.get("connect");
             if (tmpStr.equalsIgnoreCase("SSL")) {
